ROR-677 Refactor
