+++ src/main/java/no/entur/uttu/error/CodedException.java
@@ -13,17 +13,8 @@
  * limitations under the Licence.
  */
 
-package no.entur.uttu.util;
+package no.entur.uttu.error;
 
-public class CodedIllegalArgumentException extends IllegalArgumentException implements CodedException {
-    private final String code;
-
-    public CodedIllegalArgumentException(String message, String code) {
-        super(message);
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
+public interface CodedException {
+    public ErrorCodeEnumeration getCode();
 }
