+++ src/main/java/no/entur/uttu/error/ErrorCodeEnumeration.java
@@ -1,0 +1,9 @@
+package no.entur.uttu.error;
+
+public enum ErrorCodeEnumeration {
+
+    /**
+     * Organisation is not a valid operator
+     */
+    INVALID_OPERATOR
+}
