+++ src/main/java/no/entur/uttu/graphql/resource/FlexibleLinesGraphQLResource.java
@@ -82,6 +82,4 @@ public class FlexibleLinesGraphQLResource {
             Context.clear();
         }
     }
-
-
 }
