+++ src/main/java/no/entur/uttu/graphql/resource/GraphQLResourceHelper.java
@@ -25,6 +25,7 @@ import graphql.ExecutionResult;
 import graphql.GraphQL;
 import graphql.GraphQLError;
 import graphql.GraphQLException;
+import no.entur.uttu.error.CodedGraphQLError;
 import org.rutebanken.helper.organisation.NotAuthenticatedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
